chore: add reusable test harness
