+++ codebuild_specs/e2e_workflow.yml
@@ -525,13 +525,13 @@ batch:
           CLI_REGION: us-east-1
       depend-on:
         - publish_to_local_registry
-    - identifier: custom_logic_multi_gsi_update_relationships
+    - identifier: custom_logic_multi_gsi_1_record_relationships
       buildspec: codebuild_specs/run_cdk_tests.yml
       env:
         compute-type: BUILD_GENERAL1_MEDIUM
         variables:
           TEST_SUITE: >-
-            src/__tests__/custom-logic.test.ts|src/__tests__/multi-gsi-update.test.ts|src/__tests__/relationships.test.ts
+            src/__tests__/custom-logic.test.ts|src/__tests__/deploy-velocity/multi-gsi-1-record.test.ts|src/__tests__/relationships.test.ts
           CLI_REGION: us-east-2
       depend-on:
         - publish_to_local_registry
